+++ packages/app/src/features/actions/components/action-row/ActionRow.tsx
@@ -2,7 +2,7 @@ import { Token } from '@/domain/types/Token'
 import { getTokenImage } from '@/ui/assets'
 import SuccessIcon from '@/ui/assets/icons/success.svg?react'
 import WarningIcon from '@/ui/assets/icons/warning.svg?react'
-import { Button } from '@/ui/atoms/new/button/Button'
+import { Button } from '@/ui/atoms/button/Button'
 import { HorizontalScroll } from '@/ui/atoms/new/horizontal-scroll/HorizontalScroll'
 import { Tooltip, TooltipContent, TooltipTrigger } from '@/ui/atoms/tooltip/Tooltip'
 import { IconStack } from '@/ui/molecules/icon-stack/IconStack'
