+++ packages/app/src/features/dialogs/common/components/success-view/SuccessViewProceedButton.tsx
@@ -1,4 +1,4 @@
-import { Button } from '@/ui/atoms/new/button/Button'
+import { Button } from '@/ui/atoms/button/Button'
 import { ReactNode } from 'react'
 
 interface SuccessViewProceedButtonProps {
