+++ packages/app/src/features/errors/NotConnected.tsx
@@ -1,4 +1,4 @@
-import { Button } from '@/ui/atoms/new/button/Button'
+import { Button } from '@/ui/atoms/button/Button'
 import { ErrorLayout } from '@/ui/layouts/ErrorLayout'
 import { useConnectModal } from '@rainbow-me/rainbowkit'
 import { useNavigate } from 'react-router-dom'
