+++ packages/app/src/features/farm-details/components/farm-info-panel/active/ActiveFarmInfoPanel.tsx
@@ -5,8 +5,8 @@ import { Farm } from '@/domain/farms/types'
 import { USD_MOCK_TOKEN } from '@/domain/types/Token'
 import { RewardPointsSyncStatus } from '@/features/farm-details/types'
 import { getTokenImage } from '@/ui/assets'
+import { Button } from '@/ui/atoms/button/Button'
 import { DelayedComponent } from '@/ui/atoms/delayed-component/DelayedComponent'
-import { Button } from '@/ui/atoms/new/button/Button'
 import { Panel } from '@/ui/atoms/new/panel/Panel'
 import { cn } from '@/ui/utils/style'
 import { testIds } from '@/ui/utils/testIds'
