+++ packages/app/src/features/farm-details/components/tokens-to-deposit/TokensToDeposit.tsx
@@ -1,7 +1,7 @@
 import { TokenWithBalance } from '@/domain/common/types'
 import { Token } from '@/domain/types/Token'
 import { TokenCell } from '@/features/savings/components/stablecoins-in-wallet/components/TokenCell'
-import { Button } from '@/ui/atoms/new/button/Button'
+import { Button } from '@/ui/atoms/button/Button'
 import { Panel } from '@/ui/atoms/new/panel/Panel'
 import { DataTable, DataTableProps } from '@/ui/molecules/data-table/DataTable'
 import { useMemo } from 'react'
