+++ packages/app/src/features/market-details/views/components/BackNav.tsx
@@ -1,5 +1,5 @@
 import { paths } from '@/config/paths'
-import { ButtonIcon } from '@/ui/atoms/new/button/Button'
+import { ButtonIcon } from '@/ui/atoms/button/Button'
 import { LinkButton } from '@/ui/atoms/new/link-button/LinkButton'
 import { NetworkBadge } from '@/ui/atoms/new/network-badge/NetworkBadge'
 import { ArrowLeftIcon, MinusIcon } from 'lucide-react'
