+++ packages/app/src/features/my-portfolio/components/borrow-table/BorrowTable.tsx
@@ -4,7 +4,7 @@ import { OpenDialogFunction } from '@/domain/state/dialogs'
 import { borrowDialogConfig } from '@/features/dialogs/borrow/BorrowDialog'
 import { eModeDialogConfig } from '@/features/dialogs/e-mode/EModeDialog'
 import { repayDialogConfig } from '@/features/dialogs/repay/RepayDialog'
-import { Button } from '@/ui/atoms/new/button/Button'
+import { Button } from '@/ui/atoms/button/Button'
 import { Panel } from '@/ui/atoms/new/panel/Panel'
 import { ApyTooltip } from '@/ui/molecules/apy-tooltip/ApyTooltip'
 import { ActionsCell } from '@/ui/molecules/data-table/components/ActionsCell'
