+++ packages/app/src/features/my-portfolio/components/borrow-table/components/EModeIndicator.tsx
@@ -1,5 +1,5 @@
 import { EModeCategoryId } from '@/domain/e-mode/types'
-import { Button } from '@/ui/atoms/new/button/Button'
+import { Button } from '@/ui/atoms/button/Button'
 import { EModeBadge } from '@/ui/molecules/e-mode-badge/EModeBadge'
 import { testIds } from '@/ui/utils/testIds'
