+++ packages/app/src/features/my-portfolio/components/deposit-table/DepositTable.tsx
@@ -3,7 +3,7 @@ import { OpenDialogFunction } from '@/domain/state/dialogs'
 import { collateralDialogConfig } from '@/features/dialogs/collateral/CollateralDialog'
 import { depositDialogConfig } from '@/features/dialogs/deposit/DepositDialog'
 import { withdrawDialogConfig } from '@/features/dialogs/withdraw/WithdrawDialog'
-import { Button } from '@/ui/atoms/new/button/Button'
+import { Button } from '@/ui/atoms/button/Button'
 import { Panel } from '@/ui/atoms/new/panel/Panel'
 import { ApyTooltip } from '@/ui/molecules/apy-tooltip/ApyTooltip'
 import { ActionsCell } from '@/ui/molecules/data-table/components/ActionsCell'
