+++ packages/app/src/features/savings/components/stablecoins-in-wallet/StablecoinsInWallet.tsx
@@ -2,7 +2,7 @@ import { OpenDialogFunction } from '@/domain/state/dialogs'
 import { convertStablesDialogConfig } from '@/features/dialogs/convert-stables/ConvertStablesDialog'
 import { savingsDepositDialogConfig } from '@/features/dialogs/savings/deposit/SavingsDepositDialog'
 import { assets as uiAssets } from '@/ui/assets'
-import { Button } from '@/ui/atoms/new/button/Button'
+import { Button } from '@/ui/atoms/button/Button'
 import { Panel } from '@/ui/atoms/new/panel/Panel'
 import { DataTable, DataTableProps } from '@/ui/molecules/data-table/DataTable'
 import { cn } from '@/ui/utils/style'
