+++ packages/app/src/features/topbar/components/topbar-wallet/TopbarWalletButton.tsx
@@ -4,7 +4,7 @@ import { ButtonHTMLAttributes, forwardRef } from 'react'
 import { WalletDropdownTriggerInfo } from '@/features/topbar/types'
 import { shortenAddress } from '@/ui/utils/shortenAddress'
 
-import { Button, ButtonIcon } from '@/ui/atoms/new/button/Button'
+import { Button, ButtonIcon } from '@/ui/atoms/button/Button'
 import { cn } from '@/ui/utils/style'
 
 export interface TopbarWalletButtonProps extends ButtonHTMLAttributes<HTMLButtonElement>, WalletDropdownTriggerInfo {
