+++ packages/app/src/ui/layouts/app-layout/components/PageNotSupportedWarning.tsx
@@ -1,4 +1,4 @@
-import { Button } from '@/ui/atoms/new/button/Button'
+import { Button } from '@/ui/atoms/button/Button'
 import { cn } from '@/ui/utils/style'
 import { testIds } from '@/ui/utils/testIds'
 import { AlertTriangle } from 'lucide-react'
